Add common messaging attributes
Similar to the way in which we handle database attributes, provide a
mechanism that allows for centralized messaging configuration. Right
now this supports rabbit, but there is no reason other message queues
could not be added in the default['openstack']['mq'] namespace.

This bumps the version to 3.0.1 and all tests pass.

Change-Id: Id9d34fa7c256920dd575eee57c7b3820dacf8068